backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
sqladmin>=0.19.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0

